Add logic for reconnecting to Photon Server
There is no handling of the event where the client fails to connect to
Photon servers. We should add a scene that is loaded if the client fails
to connect, and an option for the user to attempt re-connection.

Let's
- create a ConnectionFailed scene, add button for retrying connection
- create ServerReconnector.cs and Reconnect() method to load the Loading
scene once button in ConnectionFailed scene is clicked
- override the default OnDisconnected method in ServerConnector and load
the ConnectionFailed scene