Restructure combat script component
Combat script component has been restructured so that combat abilities
(such as melee attack, ranged attack, block, dodge) can be added to
actors in a modular fashion. NOTE: this commit creates new versions
of the actor prefabs with a "Draft" prefix in the names. This is for
resolving prefab merge conflicts; the old prefabs should be deleted and
the draft prefabs renamed to replace them ASAP.

Let's
- make Combat class concrete instead of abstract
- create an abstract CombatAbility class with a CombatAbilityIdentifier
  enum
- create a list of combat abilities each tagged by an identifier in
  Combat, which can be filled out in the inspector
- convert child Combat classes (MeleeCombat, RangedCombat, etc.) into
  CombatAbility child classes (MeleeAttackAbility, RangedAttackAbility,
  etc.)
- create Combat Abilities child object that holds the combat ability
  script components an enter can have
- fill out the lists of combat abilities for each actor prefab