Allow for aerial actors to fall out of the sky
Toggling gravity on for actors using AirMovement will cause them to
enter a new FallingState. There is also a separate boolean,
CanLandOnGround, that can be toggled. This could be used in the future
for allowing aerial actors to land while alive, but for now toggling
gravity + setting CanLandOnGround is used when the actor dies and
crashes to the ground.

Let's
- add new Death_Falling animation for aerial actors
- add isFalling animator boolean field for aerial actors
- add FallingState and GroundedState in the AirMovementStates namespace
- set new CanLandOnGround boolean to true and toggle gravity on in the
  HandleDeathEvent method in DragonPlayerController and
  DragonEnemyController
- transition to FallingState in AirMovement.ToggleGravity