Import Cinemachine and implement basic camera movement
Add the Cinemachine package for camera movement implementation.

Add CinemachineVirtualCamera object to the game scene. Add a reference
to this camera in Player Spawner script, so when the player is spawned,
the camera is attached to them and will follow them.

Next, need to maybe resize the area and/or add camera boundaries
(currently the dragon spawns high enough that outside the background can
be seen).