Implement ranged attack for dragon enemy
The dragon enemy will continuously ready and fire a projectile. The
ready time is 4 seconds, with the first 3 seconds spent tracking the
dragon and then locking on. The final second is for the dragon to
get out of the way. The dragon enemy cannot yet shield itself and will
die when hit by a ranged projectile from the dragon.

Let's
- add remaining animations for ranged dragon enemy (ready attack,
  attack, hurt, death, jump, jump to fall, fall)
- Rename legacy ReadyChargeAttackState to ReadyRangedAttackState
- update ranged attack ability with a ready duration
- add ranged attack ability to ranged dragon enemy
- temporarily spawn ranged dragon enemy in player spawner