Implement AI for blocking ranged projectiles
The ranged dragon enemy will react to the dragon firing a projectile by
raising their guard to block the projectile (if it's on a collision
course with said enemy). Once the projectile is destroyed, they will end
their block. This is achieved by using listeners for new events, with
the bulk of the logic in AddListenerToCombatTargetRangedAttackNode.

Let's
- create a RangedProjectileEvent class in the RangedProjectile script
  file
- invoke a RangedProjectileEvent after instantiating a projectile in
  RangedAttackState
- add a UnityEvent to RangedProjectile, HitEvent, that is invoked when
  the projectile hits an object and ends its lifecycle
- implement decision-making for whether to block a projectile in a new
  AddListenerToCombatTargetRangedAttackNode
- use AddListenerToCombatTargetRangedAttackNode in the entry node of the
  ranged combat tree