Fix missing node and remove death state check
The GetVisibleCombatTargetNode was missing from the update node of the
melee combat tree, causing users of the tree to continue attacking a
target that had died due to the combat target on the blackboard not
being updated. An unnecessary check for an actor being dead in the same
node was also removed (since dead actors are now on the Dead layer).