Move projectile handler to enemy controller
The ranged attack handler in AddListenerToCombatTargetRangedAttackNode
has been moved to RangedDragonEnemyController. A new listener for ranged
attacks in Combat is used instead; the new listener in turn fires off
its own event which the handler in the enemy controller hooks into. With
this change, the behavior node can be used for any type of enemy instead
of just those that can block projectiles.