Expand level, add camera boundary and adjust lens size
To demonstrate camera movement, we expand the level on both sides by
adding floor collision boxes (without adding backgrounds for now as this
is temporary).

Set the camera lens' orthographic size to the height of the background,
and set the camera boundary to encompass the whole level.

This boundary object is in a new layer to avoid collisions with player
and enemy objects.