Fix a bug where dragon enemy enter block state incorrectly
If the Knight is positioned between the Dragon and the ranged Dragon enemy, the Dragon's projectiles would still cause the enemy to enter block state even if the projectiles would be blocked by the Knight. If the Dragon shoots projectiles in such a way that the Knight would block the projectile, the Dragon enemy should not block the attack.

This bug was caused by the Dragon's raycast not considering the Knight's layer mask. Let's fix it by including the Knight's layer mask in the raycast.