Fix knight enemy getting lost when knight mounted
The knight enemy would run on the spot below the dragon if the knight
mounted the dragon and the dragon flew out of reach. This is fixed by
adding a node in the behavior trees to check if a navigation target can
be reached before attempting to navigate there.

Let's
- add a maxReachableHeight to GroundMovement component
- use maxReachableHeight for grounded actors in a CanReachNavTargetNode
- add CanReachNavTargetNode to the behavior trees