Consume Knight buff after the Knight successfully hits an enemy
The buff given to the Knight by the Dragon will only last for one successful hit. If there are multiple ranged Dragon enemies in the level, the Dragon will have to buff the Knight each time. If the Knight attacks do not hit any enemy, the buff is not consumed.