Fix background scaling/positioning calculations
The background scaling for screen size turned out to be unecessary. With
this commit, the background is now only scaled to compensate for the
different camera sizes for the knight/dragon players. The calculation
for background positioning has also been adjusted, as it did not account
for the camera being confined when determining the offset ratios.