Fix knight stuck in jump animation when mounting
This commit fixes an issue where the knight could get stuck in its jump
animation if the dragon is mounted while it is still in its jump
animation. This also prevents the bug where, after getting stuck in its
jump animation while mounted, the knight could become trapped in its
attack animation and be unable to dismount.