Change condition for target out of melee range
This commit changes the check for an AI's combat target being out of
melee range. A target will now only be considered out of melee range if
the target is behind the AI actor. This means that the actor will not
stop readying their attack as long as their target remains ahead of
them, which currently isn't too bad since melee attackers have a short
ready time of 2 seconds.