Add duration for interactions
This commit makes each interaction take a specified amount of time to
complete. Mounting takes 0 seconds and reviving takes 3 seconds. The
player must hold down the interact button for the specified time before
the interaction completes.

A new InteractState is created and used in the combat state machine
during interaction, which is not ideal. Combat states/combat state
machine might be renamed in a future commit.

Let's
- create an InteractState as a child of CombatState
- make each Interactable take a certain amount of time
- allow interactions to be cancelled mid-interaction
- add necessary inputs to start, stop, and cancel interactions