Fix blinking bug
When knight buff is ended prematurely (eg. used on enemy), the blinking
effect still occurs. This is because if StopCoroutine is called while
ExpireBuff's WaitForSeconds is in progress, StartBlinking() is still
called. To fix this, add condition such that if RemoveBuff() is called
before blinking starts, blinking does not start.

This can also occur while blinking is in progress (because the blinking
logic also has WaitForSeconds calls). So we also add conditions to check
if RemoveBuff() has been called during blinking, and if so end the
blinking and set the knight sprite color to default.