Queue grid updates to help performance
Grid update calls from external classes are now queued up so that the
performance cost of updating the grid during runtime is spread out over
a few frames. A grid will only be updated once per frame, which should
theoretically help with performance in the case where there are many
requests to update the grid in a single frame.

Let's
- change the grid update method into a coroutine
- enqueue grid update requests and process one per frame