Clear restart message if requesting exit and vice versa
Let's:
* Change the toggleable buttons in escape menu such that only one of them can be toggled at a time.

For example if the "restart level" button is toggled and the player clicks on the "exit to role select" button, the restart button is toggled off and the exit button is toggled on. The same applies in reverse. This is to avoid the player from requesting a restart and exit at the same time.

* Clear the requesting restart message if the player then switches to an exit request and vice versa. A check is performed to ensure that the player does not wrongly clear his partner's message.
* Clear the partner's restart request in RoomManager if they request an exit and vice versa.