Mark node as walkable if obstacle is in 1 quadrant
A node will be considered walkable if an obstacle that occupies part of
it only occupies one quadrant of the node (assuming the node is a square
box). This fixes an issue where a navigation target could be up against
a wall in an unwalkable node, causing pathfinding to fail.

This commit also fixes an issue with incorrect getting node from a world
point on the Y axis.