Add visual cues for lock-on attacks
Let's:
* Add fields for colour and width increase for ProjectilePathDisplay.

These fields will affect the appearance of the projectile line path for ranged enemies. When a ranged enemy has locked-on to a player, the line will change colour and increase in width to give players a clear signal that the enemy is about to attack. The current colour set for both ranged enemies is red, while the current width increase set is 0.03.