Fix dragon dodging backwards
The dragon was capable of dodging backwards (i.e. moving backwards
instead of flipping around and moving forward) if the directional key
was pressed a split second before the dodge key. This has been fixed. In
fixing this, the possiblity of the dragon entering the buff tutorial
tunnel butt-first has been removed.