Move check for visible target into separate node
Checking for a visible combat target versus getting a combat target
using a radial grab have been separated into different nodes, with the
former used in the idle tree and the latter in the combat tree. This
prevents an issue where the same combat actor target could have a
listener attached to it twice by the AI targeting it.